feat: 쪽지 모달 컴포넌트 분리
