+++ src/components/common/modal/modal.ts
@@ -1,0 +1,7 @@
+export const modalTypes = {
+  login: 'login',
+  inbox: 'inbox',
+  outbox: 'outbox',
+  reply: 'reply',
+  sendNote: 'sendNote',
+};
