+++ src/components/login/SetProfile.tsx
@@ -6,7 +6,7 @@ import {
   useProfileImage,
   useUpdateProfile,
 } from 'hooks';
-import ModalNavigator from '../common/ModalNavigator';
+import ModalNavigator from '../common/modal/ModalNavigator';
 import ConfirmButton from './ConfirmButton';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { getUserInfoData, updateUserInfoData } from 'apis/mypageUsers';
