+++ src/routes/Router.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Header from 'components/Header';
-import ModalContainer from 'components/ModalContainer';
+import ModalContainer from 'components/common/modal/ModalContainer';
 
 const MainComponentPage = React.lazy(() => import('pages/MainPage'));
 const MyPageComonentPage = React.lazy(() => import('pages/MyPage'));
